requirements.txt
djangorestframework
django
channels
channels-redis
celery
redis
daphne
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
idna==2.8
multidict==4.6.1
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.7
yarl==1.4.1
tqdm
django-filter
